django/requirements.txt
Django
djangorestframework
pandas
numpy
scikit-learn
lightgbm
chardet
matplotlib
django-cors-headers
